backend/requirements.txt
pandas>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
openpyxl>=3.1.0
mlx>=0.28.0
mlx-lm>=0.19.0
Pillow>=10.0.0
huggingface-hub>=0.20.0